+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# unreleased
+
+- Adds MINITEST_SUITE_ONLY and MINITEST_SUITE_EXCEPT options
+
 # 0.0.1
 
 - initial release