+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# unreleased
+# 0.0.2
 
 - Adds Minitest::Suite.order= for specifying initial suite order
 - Adds MINITEST_SUITE_ONLY and MINITEST_SUITE_EXCEPT options